requirements.txt
absl-py==0.11.0
dataclasses==0.6
mediapipe==0.8.1
numpy==1.19.3
opencv-python==4.4.0.46
Pillow==8.0.1
protobuf==3.14.0
six==1.15.0